Кастомные операторы вынесены в отдельный модуль
